Plugin that enables Git-administrators to change PatchSet to draft
Git-administrators sometimes needs to set PatchSets as draft, to
enable users to remove PatchSets, for legal or other reasons.
Today this operation is done by setting PatchSet
to draft directly in the Review-db. This operation is troublesome
and error-prone, due to the human factor.

This plugin enables administrators to set PatchSet as draft with
a simple ssh-command.

Change-Id: Ied35804af652778742b941720b38e351673f3cff