Plugin that enables administrators to change PatchSet to draft
Administrators sometimes need to set PatchSets as draft, to enable
users to remove PatchSets, for reasons such as

- User has uploaded huge files that we don't want in the repository,
  even if only on the refs/changes namespace.

- User has uploaded proprietary or secret code that for legal reasons
  must not be stored on the Gerrit server.

Today this operation is done by setting the PatchSet to draft directly
in the review database. This operation is troublesome and error-prone,
due to the human factor.

This plugin enables administrators to set the PatchSet to draft with a
simple ssh command.

Change-Id: Id9991aa3ee9249ec371f406e84dbc1a28e0a9bb7