Update patch set 1
Patch Set 1:

(1 comment)

> Not sure if I've done this properly.  I had some trouble creating
 > the plugins-force-draft group; it intermittently failed to load the
 > page saying I did not have access, or the group did not exist.

This is "normal" and I always get it when I create a new group. I believe this is because the new group is not immediately available on all nodes and it takes some time until the new group is visible everywhere.

 > 
 > I somehow managed to add Sven, myself, and the gerrit maintainers
 > group.

Adding the maintainer group is not needed. It already has permissions on the parent project which are inherited.

Patch-set: 1