Update patch set 2
Patch Set 2:

> Yes, that's fine. I just checked the code in patchSet.isDraft: It
 > only considers the draft status of the PS

That's the only thing that decides if a patchset is a draft or not.

 > , so after (manual?)
 > manipulation one can arrive in this pathological state that a
 > non-draft PS is contained in a draft change.

Yes, after (manual) manipulation you can end up in all sorts of states and if you do this plugin not handling these edge cases is not what you need to worry about :-)
The state of non-draft ps in non-draft change is not valid. This is not the place to check if data is corrupt.

Patch-set: 2