THIS REPO IS DEPRECATED
The main function of this plugin was to allow Gerrit admins to change
the state of one or several patchsets of a change to 'draft' so that
the change owner could delete the change if f.i. the change owner had
unadvertently pushed sensetive information such as passwords.

Since v2.14 change owners can delete changes so there is no use for
this plugin and it will no longer be maintained.

Change-Id: I7ce2c050f659de050359214086b1e925d1a040c9